Update Timelineforwordcountsmodifiedfordate.png
Each graph shows the time-trend of a keyword/hashtag supporting/against Hillary Clinton for the day of October 28th.

The x-axis shows the number of hours on the day of October 28th.  "00" is 12 am, “15" is 1 pm and so on.

The Y-axis is reversed. It shows the ranking of the word usage at a given hour. For example, at 12 a.m., the hashtag "nevertrump" was the 200th most popular hashtag among all words tweeted in that hour. The higher the trend on the axis (lower number) means that a keyword is trendier.

The red dotted line shows the time of release of Comey's letter. The letter suggests reopening an investigation into Clinton's email server usage.

As can be seen from the graphs, it does look like that hashtags in support of Clinton go down after letter release and hashtags against Clinton go up!